# openapi.yaml
...
openapi: "3.0.3"
info:
  title: sc-agent
  description: API to interact with sc-agent
  version: 1.0.0
  contact:
    name: soeren schneider
    url: https://github.com/soerenschneider/sc-agent
    email: "56670304+soerenschneider@users.noreply.github.com" # this is just for linting reasons
  license:
    name: "NONE" # this is just for linting reasons
    url: "none" # this is just for linting reasons
servers:
  - url: http://localhost:9999
    description: Local server
...
paths:
...
  /v1/wol-message/{alias}:
    post:
      operationId: wolPostMessage
      summary: Send a WOL packet
      description: Craft a WOL packet that wakes up a machine using Wake-on-LAN. The alias must be provided by path.
      tags:
        - wol
      parameters:
        - name: alias
          in: path
          required: true
          schema:
            type: string
          example: notebook
          description: The configured alias of the machine to send a WOL packet to
      responses:
        '200':
          description: Machine wakeup signal sent successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '501':
          $ref: '#/components/responses/NotImplemented'
components:
  schemas:
    Problem:
      description: A generic HTTP problem
      xml:
        name: problem
        namespace: urn:ietf:rfc:7807
      properties:
        type:
          type: "string"
          description: "A URI reference that identifies the problem type"
          example: "https://example.com/probs/out-of-credit"
        title:
          type: "string"
          description: "A short, human-readable summary of the problem type"
          example: "You do not have enough credit."
        detail:
          type: "string"
          description: "A human-readable explanation specific to this occurrence of the problem"
          example: "Your current balance is 30, but that costs 50."
        instance:
          type: "string"
          description: "A URI reference that identifies the specific occurrence of the problem"
          example: "/account/12345/msgs/abc"
        status:
          type: "integer"
          description: "The HTTP status code"
          example: 400
      example:
        type: "https://example.com/probs/out-of-credit"
        title: "You do not have enough credit."
        detail: "You do not have enough credit."
        instance: "/account/12345/msgs/abc"
        status: 400
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/bad-request
            title: Bad Request
            status: 400
            detail: The request is invalid or missing required parameters.
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/bad-request
            title: Bad Request
            status: 400
            detail: The request is invalid or missing required parameters.
    Forbidden:
      description: Forbidden
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/forbidden
            title: Forbidden
            status: 403
            detail: Access is forbidden for the provided credentials.
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/forbidden
            title: Forbidden
            status: 403
            detail: Access is forbidden for the provided credentials.
    InternalServerError:
      description: Internal Server Error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/internal-server-error
            title: Internal Server Error
            status: 500
            detail: An unexpected error occurred.
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/internal-server-error
            title: Internal Server Error
            status: 500
            detail: An unexpected error occurred.
    NotFound:
      description: Not Found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/not-found
            title: Not Found
            status: 404
            detail: The requested resource was not found.
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/not-found
            title: Not Found
            status: 404
            detail: The requested resource was not found.
    NotImplemented:
      description: Not Implemented
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/not-implemented
            title: Not Implemented
            status: 501
            detail: This function is not enabled.
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/not-implemented
            title: Not Implemented
            status: 501
            detail: This function is not enabled.
...